[GTNAP-242] Tabstops refactor.
